+++ apps/renterd/components/Contracts/ContractsBatchMenu/index.tsx
@@ -1,0 +1,13 @@
+import { MultiSelectionMenu } from '@siafoundation/design-system'
+import { useContracts } from '../../../contexts/contracts'
+import { ContractsBatchDelete } from './ContractsBatchDelete'
+
+export function ContractsBatchMenu() {
+  const { multiSelect } = useContracts()
+
+  return (
+    <MultiSelectionMenu multiSelect={multiSelect} entityWord="contract">
+      <ContractsBatchDelete />
+    </MultiSelectionMenu>
+  )
+}